app/backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
